requirements.txt
beautifulsoup4==4.12.2
lxml==4.9.3
numpy==1.26.4
Pillow==10.0.0
PyYAML==6.0.1
requests==2.31.0
selenium==4.15.0
tensorboard==2.15.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
ultralytics==8.0.200
